SpacyTrainer/requirements.txt
falcon
spacy-nightly==2.0.0rc2
pytest
gunicorn